+++ include/rain/type.hpp
@@ -1,0 +1,15 @@
+/*
+Compatibility layer for C++ types.
+*/
+
+#pragma once
+
+#include <cstdlib>
+
+namespace Rain {
+	/*
+	errno_t is non-standard; however we define it here to maximize compatibility
+	with overloading library functions which return this.
+	*/
+	typedef int errno_t;
+}